IsWallは一応作れた

+++ TitleUIManager.cpp
@@ -26,8 +26,8 @@ void TitleUIManager::Draw()
 	Image::SetTransform(hPict_[0], title);
 	Image::Draw(hPict_[0]);
 
-	float screenWidth = Direct3D::screenWidth_;		//スクリーンの幅
-	float screenHeight = Direct3D::screenHeight_;
+	float screenWidth = (float)Direct3D::screenWidth_;		//スクリーンの幅
+	float screenHeight = (float)Direct3D::screenHeight_;
 	XMFLOAT3 size = Image::GetTextureSize(hPict_[1]);
 	Transform bg;
 	bg.scale_.x = (screenWidth / size.x);
